Add files via upload
This project not only allowed me to apply my knowledge of HTML and CSS but also helped develop my design skills. The landing page for "Taste Heaven" as a demo landing page serves as an effective marketing tool to attract customers and provide them with essential information, demonstrating the power of web design in the restaurant industry.